[autonomous] refactor: split container lifecycle route into validation/init modules (CF-024b)
startOrRestartContainer remains the orchestrator in lifecycle.ts; importers
(setup/handlers.ts, index.ts) unchanged.
