+++ src/routes/container/lifecycle.ts
@@ -4,349 +4,33 @@
  */
 import { Hono } from 'hono';
 import { getContainer } from '@cloudflare/containers';
-import type { Env, Session, SessionMode, UserPreferences, LlmKeys, DeployKeys, ContainerConfigPayload } from '../../types';
-import { createBucketIfNotExists, getOrCreateScopedR2Token } from '../../lib/r2-admin';
-import { seedGettingStartedDocs, reconcileAgentConfigs, reseedContextModePlugin } from '../../lib/r2-seed';
+import type { Env, Session, UserPreferences, LlmKeys, DeployKeys } from '../../types';
 import { resolveSessionMode, clampSessionModeToTier } from '../../lib/session-mode';
-import { getR2Config } from '../../lib/r2-config';
 import { getContainerContext, getSessionIdFromQuery, getContainerId } from '../../lib/container-helpers';
 import { AuthVariables } from '../../middleware/auth';
 import { createRateLimiter } from '../../middleware/rate-limit';
-import { AppError, ContainerError, NotFoundError, QuotaExceededError, toError, toErrorMessage } from '../../lib/error-types';
-import { getTierConfig, getUserTier, getEffectiveTier } from '../../lib/subscription';
+import { AppError, ContainerError, toError, toErrorMessage } from '../../lib/error-types';
+import { getTierConfig, getEffectiveTier } from '../../lib/subscription';
 import { isSaasModeActive } from '../../lib/onboarding';
-import { BUCKET_NAME_SETTLE_DELAY_MS, CONTAINER_ID_DISPLAY_LENGTH, getMaxSessions } from '../../lib/constants';
-import { getSessionKey, getPreferencesKey, getLlmKeysKey, getDeployKeysKey, listAllKvKeys, getSessionPrefix, getTimekeeperKey, getUtcMonthString, putSessionWithMetadata, type SessionListMetadata } from '../../lib/kv-keys';
+import { CONTAINER_ID_DISPLAY_LENGTH, getMaxSessions } from '../../lib/constants';
+import { getSessionKey, getPreferencesKey, getLlmKeysKey, getDeployKeysKey, putSessionWithMetadata } from '../../lib/kv-keys';
 import { getDefaultTabConfig } from '../../lib/agent-config';
-import { SetBucketNameBodySchema } from '../../lib/container-config-schema';
-import { containerLogger, getStoredBucketName } from './shared';
+import { containerLogger } from './shared';
 import { getContainerInternalCB } from '../../lib/circuit-breakers';
 import type { Logger } from '../../lib/logger';
 import { getAndDecrypt, getOrImportKey } from '../../lib/kv-crypto';
+import { resolveEffectiveSleepAfter, validateSessionAndCheckLimits } from './lifecycle-validation';
+import { setupR2Credentials, ensureBucketAndSeed, configureContainerDO } from './lifecycle-init';
+
+// Re-exported so existing importers (and the spec-anchored unit tests that
+// import these from './lifecycle') keep resolving them after the CF-024b split
+// into lifecycle-validation.ts and lifecycle-init.ts.
+export { resolveEffectiveSleepAfter, validateSessionAndCheckLimits } from './lifecycle-validation';
+export { setupR2Credentials, ensureBucketAndSeed, configureContainerDO } from './lifecycle-init';
 
 // ---------------------------------------------------------------------------
-// Extracted helpers
+// Orchestration
 // ---------------------------------------------------------------------------
-
-/**
- * Resolve the effective per-session idle-timeout value from the user's tier
- * and stored preference.
- *
- * REQ-SESSION-014 AC2: the "free" tier is locked to 15m regardless of any
- * stored preference; all other tiers honor the stored sleepAfter (or default
- * to 30m when no preference was ever set).
- *
- * Exported so the spec-anchored unit test in
- * src/__tests__/routes/session-sleep-timeout.test.ts can call it directly
- * without spinning up the full /api/container/start integration harness.
- */
-export function resolveEffectiveSleepAfter(
-  effectiveTier: string,
-  storedSleepAfter: string | undefined,
-): string {
-  if (effectiveTier === 'free') return '15m';
-  return storedSleepAfter || '30m';
-}
-
-/**
- * Build the JSON body for /_internal/setBucketName requests.
- * Extracted to avoid duplication between initial set and post-destroy re-set.
- *
- * sleepAfter is REQUIRED here (not defaulted). The /start route resolves the
- * effective value from `(effectiveTier === 'free' ? '15m' : preferences.sleepAfter || '30m')`
- * and passes it explicitly. A missing sleepAfter at this layer would mean the
- * resolution skipped a code path and we should fail loudly rather than ship a
- * silent '30m' default that lies to the user about their configured 2h pref.
- */
-function buildSetBucketNameBody(params: ContainerConfigPayload): string {
-  if (!params.sleepAfter) {
-    throw new Error(
-      'buildSetBucketNameBody: sleepAfter is required. The caller must resolve '
-      + 'it from user preferences before invoking. A silent default would mask '
-      + 'a real bug where the user\'s configured idle-timeout is ignored.'
-    );
-  }
-  const body = SetBucketNameBodySchema.parse({
-    bucketName: params.bucketName,
-    sessionId: params.sessionId,
-    userEmail: params.userEmail,
-    r2AccessKeyId: params.scopedCreds.accessKeyId,
-    r2SecretAccessKey: params.scopedCreds.secretAccessKey,
-    r2AccountId: params.r2Config.accountId,
-    r2Endpoint: params.r2Config.endpoint,
-    tabConfig: params.tabConfig,
-    workspaceSyncEnabled: params.workspaceSyncEnabled,
-    fastStartEnabled: params.fastStartEnabled,
-    ...(params.llmKeys?.openaiApiKey && { openaiApiKey: params.llmKeys.openaiApiKey }),
-    ...(params.llmKeys?.geminiApiKey && { geminiApiKey: params.llmKeys.geminiApiKey }),
-    // REQ-AGENT-029 AC2: a stored `null` on any of the three deploy creds is an
-    // explicit clear that must reach the container so a revoked credential is
-    // unset, not silently left stale. `undefined` stays omitted (no change); a
-    // string sets the value.
-    ...(params.deployKeys?.githubToken !== undefined && { githubToken: params.deployKeys.githubToken }),
-    ...(params.deployKeys?.cloudflareApiToken !== undefined && { cloudflareApiToken: params.deployKeys.cloudflareApiToken }),
-    ...(params.deployKeys?.cloudflareAccountId !== undefined && { cloudflareAccountId: params.deployKeys.cloudflareAccountId }),
-    ...(params.encryptionKey && { encryptionKey: params.encryptionKey }),
-    sessionMode: params.sessionMode,
-    sleepAfter: params.sleepAfter,
-    // REQ-MEM-001 AC4: forward the user's IANA timezone so the capture
-    // pipeline's TZ resolution produces wall-clock filenames matching
-    // the user's location instead of UTC.
-    ...(params.userTimezone && { userTimezone: params.userTimezone }),
-  });
-  return JSON.stringify(body);
-}
-
-/**
- * Get scoped R2 credentials for a user's bucket.
- * Wraps getOrCreateScopedR2Token with logging and error translation.
- */
-async function setupR2Credentials(
-  env: Env,
-  userEmail: string,
-  r2AccountId: string,
-  bucketName: string,
-  logger: Logger,
-  cryptoKey?: CryptoKey | null,
-): Promise<{ accessKeyId: string; secretAccessKey: string; tokenId: string }> {
-  try {
-    return await getOrCreateScopedR2Token(
-      userEmail,
-      r2AccountId,
-      env.CLOUDFLARE_API_TOKEN,
-      bucketName,
-      env.KV,
-      cryptoKey,
-    );
-  } catch (error) {
-    logger.error('Failed to create scoped R2 token', toError(error), { bucketName });
-    throw new ContainerError('r2_credentials', toErrorMessage(error));
-  }
-}
-
-/**
- * Validate that the session exists and check concurrent session limits.
- * Returns the session data if valid.
- *
- * @throws NotFoundError if session doesn't exist
- * @throws QuotaExceededError if session limit exceeded
- */
-export async function validateSessionAndCheckLimits(params: {
-  env: Env;
-  bucketName: string;
-  sessionId: string;
-  maxSessions: number;
-  subscriptionTier?: string;
-  accessTier?: string;
-  billingStatus?: string;
-  billingPeriodEnd?: string;
-}): Promise<Session> {
-  const { env, bucketName, sessionId, maxSessions, subscriptionTier, accessTier, billingStatus, billingPeriodEnd } = params;
-
-  const sessionKey = getSessionKey(bucketName, sessionId);
-  const sessionData = await env.KV.get<Session>(sessionKey, 'json');
-  if (!sessionData) {
-    throw new NotFoundError('Session', sessionId);
-  }
-
-  // Session limit + quota checks. Bypass when stress testing.
-  if (env.STRESS_TEST_MODE !== 'active') {
-    // Resolve tier once for both session limit and quota checks (cached 60s)
-    const isSaas = isSaasModeActive(env.SAAS_MODE);
-    let resolvedTier: ReturnType<typeof getUserTier> | null = null;
-    if (isSaas) {
-      try {
-        const tiers = await getTierConfig(env.KV);
-        resolvedTier = getUserTier(getEffectiveTier(subscriptionTier, accessTier, billingStatus, billingPeriodEnd), tiers);
-      } catch { /* fall back to role-based */ }
-    }
-
-    // Session limit: tier-based in SaaS mode, role-based otherwise.
-    // Uses list metadata to count running sessions (zero individual KV.get calls).
-    const effectiveMaxSessions = resolvedTier?.maxSessions ?? maxSessions;
-    const sessionKeys = await listAllKvKeys(env.KV, getSessionPrefix(bucketName));
-    // Count running sessions from authoritative KV status (the container
-    // writes 'stopped' on exit, so no read-side staleness reconciliation).
-    let runningCount = 0;
-    for (const key of sessionKeys) {
-      const rawMeta = key.metadata as SessionListMetadata | null;
-      if (rawMeta && rawMeta.s) {
-        // Fast path: read status from list metadata
-        const keySessionId = key.name.split(':').pop();
-        if (rawMeta.s === 'r' && keySessionId !== sessionId) runningCount++;
-      } else {
-        // Fallback: pre-migration key without metadata
-        const s = await env.KV.get<Session>(key.name, 'json');
-        if (!s || s.id === sessionId) continue;
-        if (s.status === 'running') runningCount++;
-      }
-    }
-
-    if (runningCount >= effectiveMaxSessions) {
-      throw new QuotaExceededError(
-        `Session limit reached (${runningCount}/${effectiveMaxSessions}). Stop an existing session to start a new one.`
-      );
-    }
-
-    // Usage quota check (SaaS mode only)
-    if (isSaas && resolvedTier && resolvedTier.monthlySeconds !== null) {
-      try {
-        const usageRecord = await env.KV.get(getTimekeeperKey(bucketName), 'json') as { thisMonth?: { month: string; seconds: number } } | null;
-        const now = new Date();
-        const currentMonth = getUtcMonthString(now);
-        const monthlySeconds = (usageRecord?.thisMonth?.month === currentMonth)
-          ? usageRecord.thisMonth.seconds : 0;
-
-        if (monthlySeconds >= resolvedTier.monthlySeconds) {
-          const usedHours = Math.round(monthlySeconds / 3600);
-          const quotaHours = Math.round(resolvedTier.monthlySeconds / 3600);
-          throw new QuotaExceededError(
-            `Monthly compute quota reached (${usedHours}h / ${quotaHours}h). Upgrade your plan.`
-          );
-        }
-      } catch (err) {
-        if (err instanceof QuotaExceededError) throw err;
-      }
-    }
-  }
-
-  return sessionData;
-}
-
-/**
- * Create the R2 bucket if needed and seed starter docs for new buckets.
- * Returns the R2 config.
- *
- * @throws ContainerError if bucket creation fails
- */
-export async function ensureBucketAndSeed(params: {
-  env: Env;
-  bucketName: string;
-  sessionMode: SessionMode;
-  contextModeEnabled?: boolean;
-  logger: Logger;
-}): Promise<{ r2Config: { accountId: string; endpoint: string } }> {
-  const { env, bucketName, sessionMode, contextModeEnabled, logger } = params;
-
-  const r2Config = await getR2Config(env);
-  const bucketResult = await createBucketIfNotExists(
-    r2Config.accountId,
-    env.CLOUDFLARE_API_TOKEN,
-    bucketName
-  );
-
-  if (!bucketResult.success) {
-    logger.error('Failed to create bucket', new Error(bucketResult.error || 'Unknown error'), { bucketName });
-    throw new ContainerError('bucket_creation', bucketResult.error);
-  }
-  logger.info('Bucket ready', { bucketName, created: bucketResult.created });
-
-  // Seed starter docs only once, when the bucket is newly created.
-  if (bucketResult.created) {
-    try {
-      const seedResult = await seedGettingStartedDocs(env, bucketName, r2Config.endpoint, { overwrite: false });
-      logger.info('Seeded initial getting-started docs', {
-        bucketName,
-        writtenCount: seedResult.written.length,
-        skippedCount: seedResult.skipped.length,
-      });
-    } catch (error) {
-      logger.warn('Failed to seed initial getting-started docs', {
-        bucketName,
-        error: toErrorMessage(error),
-      });
-    }
-
-    try {
-      const agentResult = await reconcileAgentConfigs(env, bucketName, r2Config.endpoint, sessionMode, {
-        overwrite: false,
-        cleanup: false,
-        contextModeEnabled,
-      });
-      logger.info('Seeded initial agent configs', {
-        bucketName,
-        mode: sessionMode,
-        contextModeEnabled,
-        writtenCount: agentResult.written.length,
-        skippedCount: agentResult.skipped.length,
-      });
-    } catch (error) {
-      logger.warn('Failed to seed initial agent configs', {
-        bucketName,
-        error: toErrorMessage(error),
-      });
-    }
-  }
-
-  // Always reseed the context-mode plugin subtree on every session start
-  // when the user is on the unlimited tier in advanced mode. The 3 plugin
-  // files (plugin.json, hooks.json, README.md) are Worker-authoritative -
-  // their content lives in the deployed Worker bundle and must always
-  // reflect the latest deploy. Existing buckets seeded before a manifest
-  // change (e.g. before the mcpServers block was added) would otherwise
-  // keep the stale manifest forever, since first-bucket seeding uses
-  // overwrite:false. Cost: 3 small R2 PUTs per session start.
-  if (contextModeEnabled === true) {
-    try {
-      const reseedResult = await reseedContextModePlugin(env, bucketName, r2Config.endpoint, true);
-      logger.info('Reseeded context-mode plugin subtree on session start', {
-        bucketName,
-        writtenCount: reseedResult.written.length,
-      });
-    } catch (error) {
-      logger.warn('Failed to reseed context-mode plugin subtree', {
-        bucketName,
-        error: toErrorMessage(error),
-      });
-    }
-  }
-
-  return { r2Config };
-}
-
-/**
- * Configure the container Durable Object: set bucket name, R2 creds, and preferences.
- * Returns whether the bucket name needed an update.
- *
- * @throws ContainerError if setBucketName fails on a needed update
- */
-export async function configureContainerDO(params: ContainerConfigPayload & {
-  container: { fetch: (req: Request) => Promise<Response> };
-  containerId: string;
-  logger: Logger;
-}): Promise<{ needsBucketUpdate: boolean; setBucketBody: string }> {
-  const { container, bucketName, containerId, logger } = params;
-
-  const storedBucketName = await getStoredBucketName(container, logger, containerId);
-
-  const setBucketBody = buildSetBucketNameBody(params);
-
-  const needsBucketUpdate = storedBucketName !== bucketName;
-
-  try {
-    await getContainerInternalCB(containerId).execute(() =>
-      container.fetch(
-        new Request('http://container/_internal/setBucketName', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: setBucketBody,
-        })
-      )
-    );
-    if (needsBucketUpdate) {
-      await new Promise(resolve => setTimeout(resolve, BUCKET_NAME_SETTLE_DELAY_MS));
-    }
-    logger.info('Set bucket name', { bucketName, previousBucketName: storedBucketName });
-  } catch (error) {
-    if (needsBucketUpdate) {
-      logger.error('Failed to set bucket name', toError(error));
-      throw new ContainerError('set_bucket_name', toErrorMessage(error));
-    }
-    logger.warn('Failed to store sessionId via setBucketName', { sessionId: params.sessionId });
-  }
-
-  return { needsBucketUpdate, setBucketBody };
-}
 
 /**
  * Start or restart the container based on current state.
